+++ desktopExample/src/main/kotlin/Main.kt
@@ -1,23 +1,14 @@
 package dev.snipme.desktopexample
 
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextFieldDefaults
-import androidx.compose.material3.VerticalDivider
-import androidx.compose.material3.darkColorScheme
-import androidx.compose.material3.lightColorScheme
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
+import Dropdown
+import LineNumberSwitcher
+import Samples
+import ThemeSwitcher
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.style.TextAlign
@@ -34,63 +25,50 @@ import dev.snipme.highlights.model.SyntaxThemes.useDark
 import dev.snipme.kodeview.view.CodeTextView
 import dev.snipme.kodeview.view.material3.CodeEditText
 
-private val sampleCode =
-    """
-    class Main {
-        public static void main(String[] args) {
-            int abcd = 100;
-        }
-    }
-    """.trimIndent()
-
 fun main() = application {
-    val isDarkModeState = remember { mutableStateOf(false) }
-    val isDarkMode = isDarkModeState.value
+    var isDarkMode by remember { mutableStateOf(false) }
+    var lineNumbersEnabled by remember { mutableStateOf(false) }
+    var currentLanguage by remember { mutableStateOf(SyntaxLanguage.DEFAULT) }
 
-    val highlightsState = remember {
+    var highlights by remember {
         mutableStateOf(
-            Highlights.Builder(code = sampleCode).build()
+            Highlights.Builder(code = Samples.kotlin).build()
         )
     }
-    val highlights = highlightsState.value
 
     fun updateSyntaxTheme(theme: SyntaxTheme) {
-        highlightsState.value = highlights.getBuilder()
-            .theme(theme)
-            .build()
-    }
-
-    fun updateSyntaxLanguage(language: SyntaxLanguage) {
-        highlightsState.value = highlights.getBuilder()
-            .language(language)
-            .build()
+        highlights = highlights.getBuilder().theme(theme).build()
     }
 
     MaterialTheme(colorScheme = if (isDarkMode) darkColorScheme() else lightColorScheme()) {
         Window(
             onCloseRequest = ::exitApplication,
-            title = "KodeView example",
+            title = "KodeView Example",
             state = rememberWindowState()
         ) {
             Surface {
                 Column(
-                    modifier = Modifier
-                        .fillMaxSize()
-                        .padding(16.dp),
+                    modifier = Modifier.fillMaxSize().padding(16.dp),
                     verticalArrangement = Arrangement.Center,
                 ) {
                     Spacer(Modifier.height(8.dp))
 
                     ThemeSwitcher(
                         isDarkMode,
-                        modifier = Modifier.fillMaxWidth(),
-                    ) { setToDarkMode ->
-                        isDarkModeState.value = setToDarkMode
-                        updateSyntaxTheme(highlights.getTheme().useDark(setToDarkMode)!!)
+                        modifier = Modifier.fillMaxWidth()
+                    ) { dark ->
+                        isDarkMode = dark
+                        updateSyntaxTheme(highlights.getTheme().useDark(dark)!!)
                     }
 
-
                     Spacer(Modifier.height(16.dp))
+
+                    LineNumberSwitcher(
+                        lineNumbersEnabled,
+                        modifier = Modifier.fillMaxWidth()
+                    ) { enabled ->
+                        lineNumbersEnabled = enabled
+                    }
 
                     Text(
                         modifier = Modifier.fillMaxWidth(),
@@ -99,57 +77,74 @@ fun main() = application {
                         textAlign = TextAlign.Center,
                     )
 
-                    Spacer(modifier = Modifier.size(16.dp))
+                    Spacer(Modifier.size(16.dp))
 
                     Row(
-                        modifier = Modifier
-                            .weight(1f)
-                            .fillMaxWidth(),
+                        modifier = Modifier.weight(1f).fillMaxWidth()
                     ) {
-                        CodeTextView(
-                            modifier = Modifier.weight(1f),
-                            highlights = highlights,
-                        )
+                        key(currentLanguage) {
+                            CodeTextView(
+                                modifier = Modifier.weight(1f)
+                                    .verticalScroll(rememberScrollState()),
+                                highlights = highlights.getBuilder()
+                                    .code(Samples.getSampleCode(highlights.getLanguage())).build(),
+                                showLineNumbers = lineNumbersEnabled,
+                                hasHorizontalScroll = true,
+                                textStyle = MaterialTheme.typography.bodyMedium,
+                            )
+                        }
+
                         VerticalDivider(Modifier.padding(8.dp))
-                        CodeEditText(
-                            modifier = Modifier.weight(1f),
-                            label = { Text("Edit code") },
-                            highlights = highlights,
-                            onValueChange = { textValue ->
-                                highlightsState.value = highlights.getBuilder()
-                                    .code(textValue)
-                                    .build()
-                            },
-                            colors = TextFieldDefaults.colors(
-                                unfocusedContainerColor = Color.Transparent,
-                                focusedContainerColor = Color.Transparent,
-                                focusedIndicatorColor = Color.Transparent,
-                                unfocusedIndicatorColor = Color.Transparent,
-                                disabledIndicatorColor = Color.Transparent,
-                                errorIndicatorColor = Color.Transparent,
-                            ),
-                        )
+                        key(highlights.getLanguage()) {
+                            CodeEditText(
+                                modifier = Modifier.weight(1f)
+                                    .verticalScroll(rememberScrollState()),
+                                highlights = highlights,
+                                showLineNumbers = lineNumbersEnabled,
+                                hasHorizontalScroll = true,
+                                textStyle = MaterialTheme.typography.bodyMedium,
+                                label = { Text("Edit code") },
+                                onValueChange = { newText ->
+                                    highlights = highlights.getBuilder().code(newText).build()
+                                },
+                                colors = TextFieldDefaults.colors(
+                                    unfocusedContainerColor = Color.Transparent,
+                                    focusedContainerColor = Color.Transparent,
+                                    focusedIndicatorColor = Color.Transparent,
+                                    unfocusedIndicatorColor = Color.Transparent,
+                                    disabledIndicatorColor = Color.Transparent,
+                                    errorIndicatorColor = Color.Transparent,
+                                ),
+                            )
+                        }
                     }
 
-                    Spacer(modifier = Modifier.size(16.dp))
+                    Spacer(Modifier.size(16.dp))
 
                     Dropdown(
                         options = SyntaxThemes.getNames(),
-                        selected = SyntaxThemes.themes().keys.indexOf(highlights.getTheme().key),
+                        selected = SyntaxThemes.themes().keys.indexOf(highlights.getTheme().key)
                     ) { selectedThemeName ->
                         updateSyntaxTheme(
                             SyntaxThemes.themes(isDarkMode)[selectedThemeName.lowercase()]!!
                         )
                     }
 
-                    Spacer(modifier = Modifier.size(16.dp))
+                    Spacer(Modifier.size(16.dp))
 
                     Dropdown(
                         options = SyntaxLanguage.getNames(),
                         selected = SyntaxLanguage.getNames().indexOfFirst {
                             it.equals(highlights.getLanguage().name, ignoreCase = true)
-                        }) { selectedLanguage ->
-                        updateSyntaxLanguage(SyntaxLanguage.getByName(selectedLanguage)!!)
+                        }
+                    ) { selectedLanguage ->
+                        val language = SyntaxLanguage.getByName(selectedLanguage)!!
+                        if (language == highlights.getLanguage()) return@Dropdown
+                        currentLanguage = language
+                        highlights = highlights.getBuilder()
+                            .language(language)
+                            .code(Samples.getSampleCode(language))
+                            .build()
                     }
                 }
             }