+++ exampleCommon/src/commonMain/kotlin/Dropdown.kt
@@ -1,5 +1,3 @@
-package dev.snipme.androidexample
-
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding