+++ exampleCommon/src/commonMain/kotlin/ThemeSwitcher.kt
@@ -1,5 +1,3 @@
-package dev.snipme.androidexample
-
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer