+++ kodeview/src/commonMain/kotlin/dev/snipme/kodeview/view/material3/CodeEditText.kt
@@ -1,21 +1,20 @@
 package dev.snipme.kodeview.view.material3
 
+import androidx.compose.foundation.horizontalScroll
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material3.Divider
+import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.Shape
@@ -28,6 +27,8 @@ import copySpanStyles
 import dev.snipme.highlights.DefaultHighlightsResultListener
 import dev.snipme.highlights.Highlights
 import dev.snipme.highlights.model.CodeHighlight
+import dev.snipme.kodeview.view.LineNumberWrapper
+import dev.snipme.kodeview.view.rememberTextStateWithHighlights
 import generateAnnotatedString
 import updateIndentations
 import androidx.compose.material3.LocalTextStyle as LocalTextStyle3
@@ -60,80 +61,54 @@ fun CodeEditText(
     shape: Shape = TextFieldDefaults3.shape,
     colors: TextFieldColors3 = TextFieldDefaults3.colors(),
     showLineNumbers: Boolean = false,
+    hasHorizontalScroll: Boolean = false,
     lineNumberTextStyle: TextStyle = textStyle.copy()
 ) {
-    val currentText = remember {
-        mutableStateOf(
-            TextFieldValue(
-                AnnotatedString(highlights.getCode())
+    val (currentText, updateCallback) = rememberTextStateWithHighlights(
+        highlights,
+        onValueChange = onValueChange,
+        handleIndentations = handleIndentations,
+    )
+
+    val labelPadding = label?.run { 8.dp } ?: 0.dp
+    val numbersPadding = PaddingValues(top = 16.dp + labelPadding)
+
+    Surface(
+        modifier = modifier.fillMaxWidth(),
+    ) {
+        LineNumberWrapper(
+            text = currentText.text,
+            showLineNumbers = showLineNumbers,
+            hasHorizontalScroll = hasHorizontalScroll,
+            lineNumberTextStyle = lineNumberTextStyle,
+            numbersPadding = numbersPadding,
+        ) {
+            TextField3(
+                onValueChange = updateCallback,
+                value = currentText,
+                enabled = enabled,
+                readOnly = readOnly,
+                textStyle = textStyle,
+                label = label,
+                placeholder = placeholder,
+                leadingIcon = leadingIcon,
+                trailingIcon = trailingIcon,
+                isError = isError,
+                visualTransformation = visualTransformation,
+                keyboardOptions = keyboardOptions,
+                keyboardActions = keyboardActions,
+                singleLine = singleLine,
+                maxLines = maxLines,
+                minLines = minLines,
+                interactionSource = interactionSource,
+                shape = shape,
+                colors = colors,
             )
-        )
-    }
-
-
-    LaunchedEffect(highlights) {
-        highlights.getHighlightsAsync(object : DefaultHighlightsResultListener() {
-            override fun onSuccess(result: List<CodeHighlight>) {
-                currentText.value = currentText.value.copy(
-                    annotatedString = result.generateAnnotatedString(currentText.value.text),
-                )
-            }
-        })
-    }
-
-    fun updateNewValue(change: TextFieldValue) {
-        val updated = change.updateIndentations(handleIndentations)
-        if (updated.text != currentText.value.text) {
-            onValueChange(updated.text)
         }
-
-        currentText.value = updated.copySpanStyles(
-            currentText.value
-        )
-    }
-
-    Row(modifier = modifier.fillMaxWidth()) {
-        if (showLineNumbers) {
-            val lines = currentText.value.text.lines().size.coerceAtLeast(minLines)
-            Column(
-                horizontalAlignment = Alignment.End,
-                modifier = Modifier
-                    .padding(top = 16.dp, end = 8.dp)
-            ) {
-                (1..lines).forEach { i ->
-                    Text(
-                        text = i.toString(),
-                        style = lineNumberTextStyle,
-                    )
-                }
-            }
-        }
-
-        TextField3(
-            modifier = Modifier.weight(1f),
-            onValueChange = ::updateNewValue,
-            value = currentText.value,
-            enabled = enabled,
-            readOnly = readOnly,
-            textStyle = textStyle,
-            label = label,
-            placeholder = placeholder,
-            leadingIcon = leadingIcon,
-            trailingIcon = trailingIcon,
-            isError = isError,
-            visualTransformation = visualTransformation,
-            keyboardOptions = keyboardOptions,
-            keyboardActions = keyboardActions,
-            singleLine = singleLine,
-            maxLines = maxLines,
-            minLines = minLines,
-            interactionSource = interactionSource,
-            shape = shape,
-            colors = colors,
-        )
     }
 }
 
+// TODO Refactor and test
 @Composable
 fun CodeEditTextSwiftUi(
     highlights: Highlights,
@@ -200,8 +175,7 @@ fun CodeEditTextSwiftUi(
         if (showLineNumbers) {
             val lines = currentText.value.text.lines().size.coerceAtLeast(minLines)
             Column(
-                modifier = Modifier
-                    .padding(top = 16.dp, end = 8.dp) // Align with TextField's internal padding
+                modifier
             ) {
                 (1..lines).forEach { i ->
                     Text(
@@ -212,8 +186,15 @@ fun CodeEditTextSwiftUi(
             }
         }
 
+        // TODO Extract common logic to a wrapper
+
+        val lineNumberAwareModifier = if (showLineNumbers)
+            modifier.horizontalScroll(rememberScrollState())
+        else
+            modifier
+
         TextField3(
-            modifier = Modifier.weight(1f),
+            modifier = lineNumberAwareModifier.weight(1f),
             value = currentText.value,
             onValueChange = ::updateNewValue,
             enabled = enabled,